requirements.txt
streamlit
streamlit-webrtc
numpy
soundfile
librosa
torch
torchaudio
pyannote.audio
openai
transformers
